+++ bin/easy_mvvm.dart
@@ -5,6 +5,7 @@ import 'dart:io';
 import 'package:args/command_runner.dart';
 import 'package:easy_mvvm/src/commands/create/create_command.dart';
 import 'package:easy_mvvm/src/commands/init_command.dart';
+import 'package:easy_mvvm/src/commands/version_command.dart';
 import 'package:pubspec_yaml/pubspec_yaml.dart';
 
 PubspecYaml getPubspecInfo() {
@@ -19,13 +20,47 @@ PubspecYaml getPubspecInfo() {
   }
 }
 
-void main(List<String> arguments) {
+extension on PackageDependencySpec {
+  String? version() => iswitch(
+        sdk: (p) => p.version.unsafe,
+        git: (p) => null,
+        path: (p) => null,
+        hosted: (p) => p.version.unsafe,
+      );
+}
+
+T? firstOrNull<T>(List<T> t) {
+  var iterator = t.iterator;
+  if (iterator.moveNext()) return iterator.current;
+  return null;
+}
+
+void main(List<String> arguments) async {
   CommandRunner? runner;
   try {
-    PubspecYaml pubspecYaml = getPubspecInfo();
-    runner = CommandRunner('easy_mvvm', 'Create mvvm project, services, and views with ease')
+    final PubspecYaml pubspecYaml = getPubspecInfo();
+
+    final List<PackageDependencySpec> allDependencies = [
+      ...pubspecYaml.dependencies,
+      ...pubspecYaml.dependencyOverrides,
+      ...pubspecYaml.devDependencies,
+    ];
+
+    final List<PackageDependencySpec> easyMvvms = allDependencies
+        .where((dependency) =>
+            dependency.version() != null && dependency.package() == 'easy_mvvm')
+        .toList(growable: false);
+
+    easyMvvms.sort((a, b) {
+      return (a.version() ?? '').compareTo(b.version() ?? '');
+    });
+
+    runner = CommandRunner(
+        'easy_mvvm', 'Create mvvm project, services, and views with ease')
       ..addCommand(InitCommand(packageName: pubspecYaml.name))
       ..addCommand(CreateCommand(packageName: pubspecYaml.name))
+      ..addCommand(
+          VersionCommand(easyMvvmVersion: firstOrNull(easyMvvms)?.version()))
       ..run(arguments).catchError((error) {
         if (error is UsageException && runner != null) {
           print(runner.commands[arguments.first]?.usage);
@@ -44,58 +79,4 @@ void main(List<String> arguments) {
     print(error);
     exit(2);
   }
-
-  // final ArgParser serviceParser = ArgParser()
-  //   ..addOption('singleton', abbr: 's', help: 'Creates a singleton service')
-  //   ..addOption('lazy-singleton', abbr: 'l', help: 'Creates a lazy singleton service (Recommended')
-  //   ..addOption('factory', abbr: 'f', help: 'Creates a factory for a service')
-  //   ..addFlag('async', abbr: 'a', negatable: false);
-  //
-  // final ArgParser mainParser = ArgParser()
-  //   ..addCommand('init')
-  //   ..addCommand('service', serviceParser)
-  //   ..addCommand('view')
-  //   ..addFlag('help', abbr: 'h', help: 'Output command usage', negatable: false);
-  //
-  // final ArgResults results = mainParser.parse(arguments);
-  //
-  // if (results.wasParsed('help') || results.arguments.isEmpty) {
-  //   print(mainParser.usage);
-  //   exit(0);
-  // }
-
-  // final ArgParser parser = ArgParser()
-  //   ..addOption('init', abbr: 'i', help: 'Creates the mvvm folder structure')
-  //   ..addOption('service', abbr: 's',
-  //       allowed: ['singleton', 'lazy-singleton', 'factory', ],
-  //       allowedHelp: {
-  //         'singleton': 'Creates a singleton service',
-  //         'lazy-singleton': 'Creates a lazy singleton service (Recommended)',
-  //         'factory': 'Creates a factory for a service',
-  //       },
-  //       defaultsTo: 'lazy-singleton', help: 'Creates a service')
-  //   ..addOption('view', abbr: 'v', help: 'Creates a view with its view model')
-  //   ..addFlag('help', abbr: 'h', help: 'Output command usage', negatable: false);
-  //
-  // final ArgResults results = parser.parse(arguments);
-  //
-  // if (results.wasParsed('help') || results.arguments.isEmpty) {
-  //   print(parser.usage);
-  //   exit(0);
-  // }
-  //
-  // try {
-  //   if (results.wasParsed('init')) {
-  //
-  //   } else if (results.wasParsed('service')) {
-  //     print(results.arguments);
-  //   } else if (results.wasParsed('view')) {
-  //
-  //   } else {
-  //     print(parser.usage);
-  //     exit(0);
-  //   }
-  // } catch(error) {
-  //   print(error);
-  // }
 }