+++ example/lib/ui/views/counter_view/counter_view_model.dart
@@ -1,6 +1,6 @@
 import 'package:easy_mvvm/easy_mvvm.dart';
 
-class CounterViewModel extends ViewModel {
+class CounterViewModel extends EasyViewModel {
   int _counter = 0;
   int get counter => _counter;