+++ example/lib/ui/views/home_view/home_view.dart
@@ -1,14 +1,63 @@
-import 'package:easy_mvvm/easy_mvvm.dart' as mvvm;
+import 'package:easy_mvvm/easy_mvvm.dart';
 import 'package:example/ui/views/home_view/home_view_model.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
 
-class HomeView extends mvvm.View<HomeViewModel> {
-  const HomeView({Key? key, required this.title}) : super(key: key);
+class HomeView extends EasyView<HomeViewModel> {
+  const HomeView({
+    super.key,
+    required this.title,
+  });
 
   final String title;
 
   @override
-  Widget build(BuildContext context, ThemeData theme, HomeViewModel viewModel, Widget? child) {
+  bool get canPop => false;
+
+  @override
+  PopInvokedContextCallback? get onPopInvoked => (context, didPop) {
+        showDialog<void>(
+          context: context,
+          builder: (BuildContext context) {
+            return AlertDialog(
+              title: const Text('Are you sure?'),
+              content: const Text(
+                'Are you sure you want to leave this page?',
+              ),
+              actions: <Widget>[
+                TextButton(
+                  style: TextButton.styleFrom(
+                    textStyle: Theme.of(context).textTheme.labelLarge,
+                  ),
+                  child: const Text('Nevermind'),
+                  onPressed: () {
+                    Navigator.pop(context);
+                  },
+                ),
+                TextButton(
+                  style: TextButton.styleFrom(
+                    textStyle: Theme.of(context).textTheme.labelLarge,
+                  ),
+                  child: const Text('Leave'),
+                  onPressed: () {
+                    Navigator.pop(context);
+                    SystemNavigator.pop();
+                    // Navigator.pop(context);
+                  },
+                ),
+              ],
+            );
+          },
+        );
+      };
+
+  @override
+  Widget build(
+    BuildContext context,
+    ThemeData theme,
+    HomeViewModel viewModel,
+    Widget? child,
+  ) {
     return Scaffold(
       appBar: AppBar(
         title: Text(title),