+++ lib/src/commands/create/create_view_command.dart
@@ -10,7 +10,13 @@ import 'package:recase/recase.dart';
 @internal
 class CreateViewCommand extends Command {
   CreateViewCommand({required this.packageName}) {
-    argParser.addOption('path', abbr: 'p', valueHelp: 'path_name', help: 'Use if path name (route) for web is not the same as the class name');
+    argParser.addOption(
+      'path',
+      abbr: 'p',
+      valueHelp: 'path_name',
+      help:
+          'Use if path name (route) for web is not the same as the class name',
+    );
   }
 
   @override
@@ -20,7 +26,8 @@ class CreateViewCommand extends Command {
   String get description => 'Creates the view and the view model';
 
   @override
-  String get usage => super.usage.replaceFirst('view [arguments]', 'view <ClassName>');
+  String get usage =>
+      super.usage.replaceFirst('view [arguments]', 'view <ClassName>');
 
   final String packageName;
 
@@ -28,8 +35,12 @@ class CreateViewCommand extends Command {
   void run() {
     if (argResults?.rest.isNotEmpty ?? false) {
       final String className = argResults?.rest[0].trim() ?? '';
-      final String classModelName = className.contains('View') ? '${className}Model' : '${className}ViewModel';
-      final String pathName = (argResults?.wasParsed('path') ?? false) ? (argResults?['path']) : (className.replaceAll('View', '').snakeCase);
+      final String classModelName = className.contains('View')
+          ? '${className}Model'
+          : '${className}ViewModel';
+      final String pathName = (argResults?.wasParsed('path') ?? false)
+          ? (argResults?['path'])
+          : (className.replaceAll('View', '').snakeCase);
       final String viewFileName = className.snakeCase;
       final String viewModelFileName = classModelName.snakeCase;
 
@@ -37,43 +48,45 @@ class CreateViewCommand extends Command {
 
       final File routeFile = File(routePath);
       if (routeFile.existsSync()) {
-        final String viewFilePath = 'lib/ui/views/$viewFileName/$viewFileName.dart';
-        final String viewModelFilePath = 'lib/ui/views/$viewFileName/$viewModelFileName.dart';
+        final String viewFilePath =
+            'lib/ui/views/$viewFileName/$viewFileName.dart';
+        final String viewModelFilePath =
+            'lib/ui/views/$viewFileName/$viewModelFileName.dart';
         final File viewFile = File(viewFilePath);
         final File viewModelFile = File(viewModelFilePath);
 
         if (viewFile.existsSync() || viewModelFile.existsSync()) {
           if (viewFile.existsSync()) {
-            throw Exception('File with the name $viewFileName.dart already exists at path $viewFilePath');
+            throw Exception(
+                'File with the name $viewFileName.dart already exists at path $viewFilePath');
           }
 
           if (viewModelFile.existsSync()) {
-            throw Exception('File with the name $viewModelFileName.dart already exists at path $viewModelFilePath');
+            throw Exception(
+                'File with the name $viewModelFileName.dart already exists at path $viewModelFilePath');
           }
         } else {
           final viewContentCode = Library((b) {
             b.name = className;
             b.directives.addAll([
-              Directive.import('package:easy_mvvm/easy_mvvm.dart', as: 'mvvm'),
+              Directive.import('package:easy_mvvm/easy_mvvm.dart'),
               Directive.import('package:flutter/material.dart'),
-              Directive.import('package:$packageName/ui/views/$viewFileName/$viewModelFileName.dart'),
+              Directive.import(
+                  'package:$packageName/ui/views/$viewFileName/$viewModelFileName.dart'),
             ]);
             b.body.add(Class((c) {
               c.name = className;
-              c.extend = refer('mvvm.View<$classModelName>');
+              c.extend = refer('EasyView<$classModelName>');
               c.constructors.addAll([
                 Constructor((d) {
                   d.optionalParameters.addAll([
                     Parameter((e) {
+                      e.toSuper = true;
                       e.name = 'key';
                       e.named = true;
-                      e.type = refer('Key?');
                     }),
                   ]);
                   d.constant = true;
-                  d.initializers.addAll([
-                    const Code('super(key: key)'),
-                  ]);
                 }),
               ]);
               c.methods.addAll([
@@ -116,10 +129,13 @@ class CreateViewCommand extends Command {
             }));
           });
 
-          final DartEmitter emitter = DartEmitter.scoped(useNullSafetySyntax: true);
-          String viewFileContents = DartFormatter().format('${viewContentCode.accept(emitter)}');
+          final DartEmitter emitter =
+              DartEmitter.scoped(useNullSafetySyntax: true);
+          String viewFileContents =
+              DartFormatter().format('${viewContentCode.accept(emitter)}');
           // remove first line with the library import
-          viewFileContents = DartFormatter().format(viewFileContents.substring(viewFileContents.indexOf('\n') + 1));
+          viewFileContents = DartFormatter().format(
+              viewFileContents.substring(viewFileContents.indexOf('\n') + 1));
 
           // create the view file
           viewFile.createSync(recursive: true);
@@ -128,17 +144,19 @@ class CreateViewCommand extends Command {
           final viewModelContentCode = Library((b) {
             b.name = classModelName;
             b.directives.addAll([
-              Directive.import('package:easy_mvvm/easy_mvvm.dart', as: 'mvvm'),
+              Directive.import('package:easy_mvvm/easy_mvvm.dart'),
             ]);
             b.body.add(Class((c) {
               c.name = classModelName;
-              c.extend = refer('mvvm.ViewModel');
+              c.extend = refer('EasyViewModel');
             }));
           });
 
-          String viewModelFileContents = DartFormatter().format('${viewModelContentCode.accept(emitter)}');
+          String viewModelFileContents =
+              DartFormatter().format('${viewModelContentCode.accept(emitter)}');
           // remove first line with the library import
-          viewModelFileContents = DartFormatter().format(viewModelFileContents.substring(viewModelFileContents.indexOf('\n') + 1));
+          viewModelFileContents = DartFormatter().format(viewModelFileContents
+              .substring(viewModelFileContents.indexOf('\n') + 1));
 
           // create the view model file
           viewModelFile.createSync(recursive: true);
@@ -146,19 +164,25 @@ class CreateViewCommand extends Command {
 
           // modify the routes.dart file
           final List<String> routesFileLines = routeFile.readAsLinesSync();
-          routesFileLines.insert(0, 'import \'package:$packageName/ui/views/$viewFileName/$viewFileName.dart\';');
+          routesFileLines.insert(0,
+              'import \'package:$packageName/ui/views/$viewFileName/$viewFileName.dart\';');
 
-          final int classEndIndex = routesFileLines.lastIndexWhere((e) => e.contains(RegExp(r'^\}$')));
+          final int classEndIndex = routesFileLines
+              .lastIndexWhere((e) => e.contains(RegExp(r'^\}$')));
           if (classEndIndex != -1) {
-            routesFileLines.insert(classEndIndex, 'router.defineRoute<$className>(instance: () => const $className(), path: \'$pathName\', className: \'$className\');');
+            routesFileLines.insert(classEndIndex,
+                'router.defineRoute<$className>(instance: () => const $className(), path: \'$pathName\', className: \'$className\');');
 
             final String routesFileAsStringOut = routesFileLines.join('\n');
-            final String routesFileContents = DartFormatter().format(routesFileAsStringOut);
-            routeFile.writeAsStringSync(routesFileContents, mode: FileMode.write);
+            final String routesFileContents =
+                DartFormatter().format(routesFileAsStringOut);
+            routeFile.writeAsStringSync(routesFileContents,
+                mode: FileMode.write);
           }
         }
       } else {
-        final Colorize fileDoesNotExist = Colorize('lib/app/routes.dart file does not exist')..red();
+        final Colorize fileDoesNotExist =
+            Colorize('lib/app/routes.dart file does not exist')..red();
         stdout.writeln(fileDoesNotExist);
       }
     } else {