+++ lib/src/commands/version_command.dart
@@ -1,0 +1,25 @@
+import 'package:args/command_runner.dart';
+import 'package:meta/meta.dart';
+
+@internal
+class VersionCommand extends Command {
+  VersionCommand({required this.easyMvvmVersion});
+
+  final String? easyMvvmVersion;
+
+  @override
+  String get name => 'version';
+
+  @override
+  List<String> get aliases => ['v'];
+
+  @override
+  String get description => 'Print the version';
+
+  @override
+  void run() {
+    if (easyMvvmVersion?.isNotEmpty == true) {
+      print('easy_mvvm $easyMvvmVersion');
+    }
+  }
+}